+++ packages/core/icon/src/metadata.js
@@ -1381,6 +1381,11 @@ export default {
     componentName: 'PreferencesIcon',
     package: '@atlaskit/icon/glyph/preferences',
   },
+  premium: {
+    keywords: ['premium', 'icon', 'core'],
+    componentName: 'PremiumIcon',
+    package: '@atlaskit/icon/glyph/premium',
+  },
   'presence-active': {
     keywords: ['presence-active', 'presenceactive', 'dot', 'icon', 'core'],
     componentName: 'PresenceActiveIcon',