+++ packages/editor/adf-schema/src/schema/nodes/tableNodes.ts
@@ -19,6 +19,15 @@ import {
   Y75,
 } from '../../utils/colors';
 import { TableCellContent } from './doc';
+import {
+  StrongDefinition as Strong,
+  EmDefinition as Em,
+  CodeDefinition as Code,
+  StrikeDefinition as Strike,
+  SubSupDefinition as SubSup,
+  UnderlineDefinition as Underline,
+  TextColorDefinition as TextColor,
+} from '../marks';
 
 const akEditorTableNumberColumnWidth = 42;
 const DEFAULT_TABLE_HEADER_CELL_BACKGROUND = N20.toLocaleLowerCase();
@@ -34,6 +43,12 @@ const getCellAttrs = (dom: HTMLElement, defaultValues: CellAttributes = {}) => {
   if (backgroundColor && isRgb(backgroundColor)) {
     backgroundColor = rgbToHex(backgroundColor);
   }
+  const dataDefaultMarks = dom.getAttribute('data-default-marks');
+  let defaultMarks = null;
+
+  try {
+    defaultMarks = dataDefaultMarks ? JSON.parse(dataDefaultMarks) : null;
+  } catch (e) {}
 
   return {
     colspan,
@@ -43,6 +58,7 @@ const getCellAttrs = (dom: HTMLElement, defaultValues: CellAttributes = {}) => {
       backgroundColor && backgroundColor !== defaultValues['background']
         ? backgroundColor
         : null,
+    defaultMarks,
   };
 };
 
@@ -52,6 +68,7 @@ export const setCellAttrs = (node: PmNode, cell?: HTMLElement) => {
     rowspan?: number;
     style?: string;
     'data-colwidth'?: string;
+    'data-default-marks'?: string;
   } = {};
   const colspan = cell ? parseInt(cell.getAttribute('colspan') || '1', 10) : 1;
   const rowspan = cell ? parseInt(cell.getAttribute('rowspan') || '1', 10) : 1;
@@ -86,6 +103,10 @@ export const setCellAttrs = (node: PmNode, cell?: HTMLElement) => {
 
       attrs.style = `${attrs.style || ''}background-color: ${color};`;
     }
+  }
+
+  if (node.attrs.defaultMarks) {
+    attrs['data-default-marks'] = JSON.stringify(node.attrs.defaultMarks);
   }
 
   return attrs;
@@ -200,6 +221,13 @@ export interface CellAttributes {
   rowspan?: number;
   colwidth?: number[];
   background?: string;
+  /**
+   * @stage 0
+   * @forceContentValidation true
+   */
+  defaultMarks?: Array<
+    Em | Strong | Code | Strike | SubSup | Underline | TextColor
+  >;
 }
 
 // TODO: Fix any, potential issue. ED-5048
@@ -258,6 +286,7 @@ const cellAttrs = {
   rowspan: { default: 1 },
   colwidth: { default: null },
   background: { default: null },
+  defaultMarks: { default: null },
 };
 
 export const tableCell = {