+++ packages/editor/editor-core/src/plugins/analytics/types/enums.ts
@@ -56,6 +56,7 @@ export const enum ACTION_SUBJECT {
 export const enum ACTION_SUBJECT_ID {
   ACTION = 'action',
   ANNOTATE_BUTTON = 'annotateButton',
+  BLOCK_QUOTE = 'blockQuote',
   BUTTON_HELP = 'helpButton',
   BUTTON_FEEDBACK = 'feedbackButton',
   CANCEL = 'cancel',