+++ packages/editor/renderer/src/__tests__/unit/email/index.ts
@@ -16,6 +16,8 @@ import * as blockCards from '../../__fixtures__/block-cards.adf.json';
 import * as inlineCards from '../../__fixtures__/inline-cards.adf.json';
 import * as status from '../../__fixtures__/status.adf.json';
 import * as tableNumberedColumn from '../../__fixtures__/table-numbered-column.adf.json';
+import * as layoutColumnSection from '../../__fixtures__/layout-column-section.adf.json';
+import * as extensions from '../../__fixtures__/extensions.adf.json';
 
 import * as image from '../../__fixtures__/image.adf.json';
 import * as placeholder from '../../__fixtures__/placeholder.adf.json';
@@ -130,6 +132,16 @@ describe('Renderer - EmailSerializer', () => {
 
   it('should render numbered column for table', () => {
     const output = render(tableNumberedColumn);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render layout column and sections', () => {
+    const output = render(layoutColumnSection);
+    expect(output).toMatchSnapshot();
+  });
+
+  it('should render extension placeholders', () => {
+    const output = render(extensions);
     expect(output).toMatchSnapshot();
   });
 });