+++ packages/editor/renderer/src/email/nodes/paragraph.ts
@@ -1,17 +1,19 @@
-import { applyMarks, createTag, serializeStyle } from '../util';
+import { applyMarks, createTag } from '../util';
 import { NodeSerializerOpts } from '../interfaces';
 
-const css = serializeStyle({
-  margin: '0',
-  'margin-bottom': '7px',
-  padding: '0px',
-  'padding-top': '7px',
-  'mso-line-height-rule': 'exactly',
-  'line-height': '24px',
-  'font-size': '14px',
-});
+export const paragraphStyles = `
+p {
+  margin: 0;
+  margin-bottom: 7px;
+  padding: 0px;
+  padding-top: 7px;
+  mso-line-height-rule: exactly;
+  line-height: 24px;
+  font-size: 14px;
+}
+`;
 
 export default function paragraph({ text, marks }: NodeSerializerOpts) {
-  const paragraph = createTag('p', { style: css }, text);
+  const paragraph = createTag('p', {}, text);
   return applyMarks(marks, paragraph);
 }