+++ packages/elements/task-decision/src/__tests__/unit/ssr/samples/_decision-inline.tsx
@@ -1,11 +1,7 @@
 import * as React from 'react';
 import DecisionItem from '../../../../components/DecisionItem';
 import DecisionList from '../../../../components/DecisionList';
-import {
-  MessageContainer,
-  dumpRef,
-  getParticipants,
-} from '../../../../../example-helpers/story-utils';
+import { dumpRef } from '../../../../../example-helpers/story-utils';
 
 export default () => (
   <div>
@@ -32,27 +28,19 @@ export default () => (
       Simple DecisionItem with 1 participant, inline (shouldn't render
       participants)
     </h3>
-    <MessageContainer>
-      <DecisionItem
-        contentRef={dumpRef}
-        participants={getParticipants(1)}
-        appearance="inline"
-      >
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua.
-      </DecisionItem>
-    </MessageContainer>
+    <DecisionItem contentRef={dumpRef} appearance="inline">
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+      tempor incididunt ut labore et dolore magna aliqua.
+    </DecisionItem>
 
     <h3>Simple DecisionList</h3>
-    <MessageContainer>
-      <DecisionList>
-        <DecisionItem contentRef={dumpRef}>
-          Hello <b>world</b>.
-        </DecisionItem>
-        <DecisionItem contentRef={dumpRef}>
-          OMG <b>YAY</b>.
-        </DecisionItem>
-      </DecisionList>
-    </MessageContainer>
+    <DecisionList>
+      <DecisionItem contentRef={dumpRef}>
+        Hello <b>world</b>.
+      </DecisionItem>
+      <DecisionItem contentRef={dumpRef}>
+        OMG <b>YAY</b>.
+      </DecisionItem>
+    </DecisionList>
   </div>
 );